fix: Squash migrations and fix entity mappings

+++ backend/mercatus/src/main/java/com/scientia/mercatus/dto/AddToCartRequestDto.java
@@ -2,5 +2,5 @@ package com.scientia.mercatus.dto;
 
 import java.math.BigDecimal;
 
-public record AddToCartRequestDto(Long productId, BigDecimal quantity) {
+public record AddToCartRequestDto(Long productId, Integer quantity) {
 }
